pipeline-manager: split program-specific DB calls into a separate module
